Renamed pinned filter toggle component to just filter toggle.
